+++ libraries/src/constants.js
@@ -8,5 +8,6 @@ export let constants = {
     url: "https://microart.cf/showroom/menuify",
     splash: splash.raw,
     regkey: "HKCR/SystemFileAssociations",
+    directoryRegkey: "HKCR/Directory",
     ffmpegReleasesEndpoint: "https://api.github.com/repos/BtbN/FFmpeg-Builds/releases/latest"
 }