+++ libraries/src/pages/editMenu.js
@@ -14,7 +14,7 @@ export async function editMenu({ inquirer }) {
         message: "Select a project to edit:",
         choices: projects.map(p => {
             return {
-                name: `${p.name} [${p.ext}] - ${p.description}`,
+                name: `${p.name} [${p.ext === "[dir]" ? chalk.magenta("directory") : p.ext}] - ${p.description}`,
                 value: p
             }
         })