+++ libraries/changelog.md
@@ -2,6 +2,14 @@
 
 ### 0.2-alpha
 
+- Added support for all file syntax (*)
+- Added support for powershell scripts
+- Added support for C# scripts
+- Added support for javascript scripts
+- Added support for pre-compiled binaries
+- Added certificate and code-signing support
+- Youtube-Dl support soon
+
 ### 0.1-alpha
 
 - Initial release